// src/models/Contact.ts
export interface Contact {
  id?: number;
  firstName: string;
  lastName: string;
  company: string;
  jobTitle: string;
  category: string;
  emailAddress: string;
  businessPhone: string;
  homePhone: string;
  mobilePhone: string;
  postalCode: string;
  faxNumber: string;
  notes: string;
  address: string;
  city: string;
  country: string;
  webPage: string;
}